sys/vfs/hammer: Temporary fix for kernel panic on volume-del
This is a temporary fix for a known race issue mentioned in
http://lists.dragonflybsd.org/pipermail/kernel/2015-August/175027.html

It may be replaced with a better alternative in the future.

The next commit also needs this, otherwise it's not guaranteed
that ondisk buffers are written back to the backing storage.

(cherry picked from commit 2f1df9cece2c18f660b07748af682ac3c47cc50f)