sbin/hammer: Make hammer info print rootvol
The existence of "root volume" in HAMMER is mentioned in
error messages and dmesg as well as hammer(8) manpage,
but the problem is that no command really tells you which
one is the root volume. This is too obvious if you have
looked at the source, but no idea if otherwise.

Currently the only one explicitly tells you seems to be
a printf in newfs_hammer added by dae8f186 in 2015.

This commit adds hammer info a printf for root volume #,
which is still not clear enough since it's just a number,
but not a block device path.