drm/i915: Avoid dead-locking the userland.
Was happening for some time and the cause is still unclear.
On exit Xorg doesn't return from kernel and is unkillable.
Sometimes even panics on haswell, it is better to avoid that.

Also reduce spam on laptops with eDP panels on i915 load.
Great opportunity to use just implemented DRM_ERROR_RATELIMITED()
macro that uses krateprintf().