Revert "<stdio.h>/<wchar.h>: Reduce namespace pollution in <wchar.h>."
This reverts commit 69e6516956c27905b4070759b0c597e6ab8b1ba5.

It looks like something in our C++ headers causes breakage in some
dports when <wchar.h> doesn't also include <stdio.h>, specifically,
including <iostream> does not seem to bring in <stdio.h>
automatically, as it is on FreeBSD and Linux.

Since it is legal (per standard) for <wchar.h> to also provide the
symbols of <stdio.h>, revert this change for now.