rescue image: Strip lvm(8).
It's built separately outside of the crunchgen'd stuff, for reasons
I'm yet not sure of.

So of course it also needs to be stripped separately because if -g is
added to the user's buildworld flags, it will go into the initrd with
debug symbols (filling it to 95%, opposed to the (currently) 73% when
not stripped).