Remove am-utils, the Berkeley automounter suite (amd, amq, etc.)
We recently got FreeBSD's autofs(5) which replaces it. FreeBSD
added notes to their am-utils and related manual pages saying
that it is obsolete and advises to use autofs(5) instead.

DragonFly's port of it is almost surely broken and the last time
I heard from someone trying to get it to work was in 2013 and
back then it just hung (in select(), according to my notes).

So I don't think removing instead of trying to fix it will do
any harm.