Run make depend in quickkernel, too.
It is much cleaner to do that, just like it is run in quickworld, too.

At the price of a small increase in build time, quickkernel will now
continue working when a new kernel header is added, which broke it
before this commit because the header would not be copied to the right
place in /usr/obj.