Switch base to use private LibreSSL libaries
All base users of the OpenSSL libraries (libssl.so and libcrypto.so) have
been modified to link against the new LibreSSL versions instead.

The OpenSSL libraries are still built by default, but nothing in base will
use them.  However, its certain that a myriad of dports link to OpenSSL
although this may change in the future.

The OpenSSL library building can be suppressed by putting NO_OPENSSL in
/etc/make.conf.  The existing OpenSSL libraries and headers are not (yet)
removed, however, even with that setting.