sbin/hammer: Remove meaningless sscanf for "@@PFS%d" format
This isn't necessary as userspace never reads "@@PFS%d" via readlink(2).
"@@%jx:%d" covers all the necessary patterns.

It also shouldn't be here because the first stat(2) call in this
function ignores the result due to possible slave PFS (before once
synced with a master PFS), so an arbitrary string that matches this
"@@PFS%d" pattern for path argument (whether it really exists or not)
makes this function return non negative fd while it should return -1.

If "@@PFS%d" is the name of an existing file or a directory, then
that's also not what this function is looking for, so it can still
be ignored.