gcc47: Suppress last codegen warning.
Turns out gcc(compilation as C) differentiates when initializing global variables that
are declared 'extern' on the same line (removal of extern attribute is ok, includes
c-common.h header that properly marks c_family_lang_mask global variable as extern).

Note: do not apply same change to gcc50 since it is compiled as C++ that is not affected.