sbin/hammer: Move hammer_btree_cmp() to cmd_show.c
This function isn't necessary in userspace unless a command wants
to walk through the B-Tree. The only hammer command that needs to
do this is hammer show, so make it a static function of hammer show
just like other node/elm related functions are static.

This is a copy-pasted function from sys/vfs/hammer/hammer_btree.c,
but keep it copy-pasted code without having it as an static inline
function in hammer_btree.h, as it has bunch of conditionals that
make asm of a caller unreadable.