sbin/hammer: Partly revert 2c1d3cef in 2010
This is the same as d6c59c72, so read d6c59c72 for details.

The code in 2c1d3cef in 2010 was later replaced by libhammer, so what
this commit actually does is revert a part of those libhammer commits,
but whatever it is, a certain directory being null mounted on somewhere
else is beyond the scope of hammer or any other filesystems.
This command originally had no such extra info, and that was better.

It's also confusing in the sense that it silently changes the output
depending on whether it's mounted or not.