Expose ld.bfd and ld.gold
It's now expected that ld, ld.bfd, and ld.gold are all installed by
modern binutils.  The -fuse-gold option on GCC depends on it (both from
base and ports) and possibly clang as well.  This work is 6 months old,
but it was never committed as collect2 utility was only recently
built and installed.

Discussed with:	dillon (6 months ago).