binutils225: Add plugin support in ar, nm, ranlib.
This finally adds support for --plugin [libplugin.so] flag usage.

Default loadable plugins should be symlinked to binutils private
/usr/libexec/binutils225/bfd-plugins/ directory.
For now that directory is not created by mtree and gcc does not
install symlink to its private liblto_plugin.so.

TBD if we want this implicit LTO support without explicit --plugin.
Note that bfd plugin enabled utilities will try to load *anything*
residing in that directory, even hangman(6)!
Still, support for it is useful for local experimentation since this
allows to test even LLVMgold.so without lots of edits in Makefiles.