binutils225: Symlink liblto_plugin.so from default base gcc compiler.
For now just for gcc50 cause there is little point for doing it on gcc47.
Also make ar, nm, ranlib utilities dynamic to avoid "Service unavailable".

Warning: even if is it tempting to go and add -flto to CFLAGS now, first
make sure that you are familiar with how -flto actually works!