sbin/hammer: Use dirname(3) instead of getdir() in scanpfsid()
scanpfsid() can use dirname(3) for fd instead of getdir().
It's actually better to use dirname(3) which strips trailing /
first and then return a pointer to directory path, instead of
getdir() which only strips trailing / if any, because that's
what scanpfsid() is trying to do with fd.

If path is
    path/to/pfs -> @@...
scanpfsid() wants to open("path/to"), and if path is
    path/to/pfs/
which is @@.. itself if the PFS is not destroyed,
scanpfsid() still wants to open("path/to").

The second one doesn't occur since 8b640662 changed getpfs()
to strip trailing / before calling scanpfsid().
However the next commit is going to partly revert 8b640662,
and this change makes the next 2 commits easier to implement,
while being logically separated topic.