sys/vfs/hammer: Mention ondisk fields are in little endian
Explicitly mention HAMMER1 uses little endian for ondisk structures,
but not host endian (which is little on currently supported arch).

This is easier because adding big endian support doesn't break
anything at least on DragonFly/x86_64.

This doesn't affect anything/anyone in reality because HAMMER1 on
DragonFly is the only HAMMER1 implementation (that I know of other
than the read only FUSE port).