machine/endian.h: Mark bswap functions always inlineable.
This solves many cases where compiler decides to make them as a local
static functions to prevent code size growth and causing -Winline.
Also then it leads to less optimal code generation.
To prevent all of this just mark them as always inlineable.
These functions are intended to be inlined and should be thought as macros.