sbin/hammer: Mention bug in hammer recover command
newfs_hammer overwrites layer1/2, but doesn't overwrite data in
blockmap beyond layer1/2, so old data is still there if the volume
has been re-formatted until the new fs overwrites old fs by using
more disk space than the old fs.

This is usually not a problem because the new fs doesn't have a
pointer to data in old fs, but hammer recover just iterates zone-2
offset till it hits end of the disk space which may include data
from the old fs. As mentioned in the comment, it's likely to hit
assertion if the old data located near the end of the volume is
B-Tree nodes depending on situation of the old fs.