sys/vfs/hammer: Rename mem_area_size to memory_log_size
The name of this field has been "memory area size" ever since it
was added in 90ecab35 in 2009, but rename it to "memory log size"
since this is the real name for this field.
(see stdout of newfs_hammer and newfs_hammer(8))

This commit changes a name of ioctl structure field, but doesn't
affect or break userspace, as the ioctl is only used or usable by
hammer volume-add command.