sbin/newfs_hammer: Fix source for filesystem info after formatting
This commit changes source of outputs after formatting a fs,
but doesn't change outputs itself, as it should result the same.
(the only difference in when -d is specified for junk space)

Now outputs are based on ondisk fields that are to be written
back to backing storage, rather than global variables within
sbin/newfs_hammer/newfs_hammer.c.

It's better to use actual ondisk fields than other variables
to make it robust against possible changes in newfs_hammer in
the future (i.e. ondisk fields are always ondisk fields).