Fix world build in NO_ALTBINUTILS edge case
In the case that the machine has been updated within 30 days but with
NO_ALTBINUTILS set, the world build fails.  This is due because the
logic to fallback to earlier binutils versions fails due to empty
directories that are installed regardless of the NO_ALBINUTILS setting.
The logic was updated to search for binutils programs rather than
directories.  In the edge case, the oldest version of binutils on the
system is used to build the native versions during the early build phases.