libc: Add temporary (get|set)affinity DF404.0 symbols
The sched_getaffinity, sched_setaffinity and related work had the new
symbols erroneously given DF404.0 versioning.  This was corrected by
commit 6262c9d, but that in turn broke any userland programs built
with the original symbol versions.

This patch redefines the old symbols while maintaining the correct
DF408.0 versions, but it is intended to be temporary.  After the
branch (e.g. Release 4.8), this commit should be reverted so that
the next development trunk won't have the DF404.0 versions.  It's
expected that all ports and third-party software is rebuilt when
upgrading to a new release.  This serves as a reasonable bridge
to allow affected users to do that rebuild.

This patch is particularly intrusive regarding the syscall assembly
file creation, so it's not something we want to leave in place
permanently in my opinion.